Extract ApiData and ThemeData into own modules for testability
